+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Added
 
 * add `Placements` [add placements support](https://github.com/mesilov/bitrix24-php-sdk/issues/274)
+* add `ApplicationStatus` with application status codes description
+* add fabric method `AccessToken::initFromPlacementRequest` when application init form placement request
 
 ### Changed