+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * add `Placements` [add placements support](https://github.com/mesilov/bitrix24-php-sdk/issues/274)
 * add `ApplicationStatus` with application status codes description
 * add fabric method `AccessToken::initFromPlacementRequest` when application init form placement request
+* add fabric method `ApplicationProfile::initFromArray` when application profile stored in ENV-variables
 
 ### Changed